requirements.txt
fastmcp
requests
python-dotenv

